Add module file
