+++ JSON/authors.json
@@ -1,0 +1,1 @@
+[{"first_name":"jack","last_name":"adams"}]
