+++ JSON/games.json
@@ -1,0 +1,1 @@
+[{"publish_date":"2020-02-02","multiplayer":true,"last_played_date":"2022-1-01"}]
